requirements.txt
altair==5.0.1
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.0
colormath==3.0.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
gymnasium==0.27.0
gymnasium-notices==0.0.1
idna==3.4
importlib-metadata==6.6.0
ipython==8.8.0
jax-jumpy==0.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
networkx==3.1
numpy==1.24.3
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipi==1.0.1
prompt-toolkit==3.0.36
protobuf==4.23.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pydeck==0.8.1b0
pygame==2.1.3.dev8
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rich==13.4.1
seaborn==0.12.2
Shimmy==0.2.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.23.1
streamlit-image-coordinates==0.1.3
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.15.0
